feat: rename @types-v5 to @types
