+++ server/src/graphql/queries/findAllFlat.ts
@@ -1,6 +1,6 @@
 import { isEmpty } from 'lodash';
-import { CoreStrapi } from '../../@types-v5';
-import { Nexus } from '../../@types-v5/graphql';
+import { CoreStrapi } from '../../@types';
+import { Nexus } from '../../@types/graphql';
 import { flatInput } from '../../controllers/utils/parsers';
 import { getModelUid } from '../../repositories/utils';
 import { getPluginService } from '../../utils/getPluginService';
