+++ server/src/graphql/queries/index.ts
@@ -1,5 +1,5 @@
-import { CoreStrapi } from '../../@types-v5';
-import { Nexus } from '../../@types-v5/graphql';
+import { CoreStrapi } from '../../@types';
+import { Nexus } from '../../@types/graphql';
 import findAllFlat from './findAllFlat';
 import findAllInHierarchy from './findAllInHierarchy';
 import findAllPerAuthor from './findAllPerAuthor';
