+++ server/src/graphql/types/getComment.ts
@@ -1,4 +1,4 @@
-import { Nexus } from '../../@types-v5/graphql';
+import { Nexus } from '../../@types/graphql';
 
 export const getComment = (nexus: Nexus) => {
   return nexus.objectType({
