+++ server/src/graphql/types/getIdentityAuthor.ts
@@ -1,4 +1,4 @@
-import { Nexus } from "../../@types-v5/graphql";
+import { Nexus } from '../../@types/graphql';
 
 export const getIdentityAuthor = (nexus: Nexus) => {
     return nexus.inputObjectType({
