+++ server/src/graphql/types/getRemoveComment.ts
@@ -1,4 +1,4 @@
-import { Nexus } from "../../@types-v5/graphql";
+import { Nexus } from '../../@types/graphql';
 
 export const getRemoveComment = (nexus: Nexus) => {
     return nexus.inputObjectType({
