+++ server/src/graphql/types/getResponsePagination.ts
@@ -1,4 +1,4 @@
-import { Nexus } from "../../@types-v5/graphql";
+import { Nexus } from '../../@types/graphql';
 
 export const getResponsePagination = (nexus: Nexus) => {
     return nexus.objectType({
