+++ server/src/routes/admin.routes.ts
@@ -1,4 +1,4 @@
-import { StrapiRoute } from '../@types-v5';
+import { StrapiRoute } from '../@types';
 
 const adminRoutes: StrapiRoute<'admin'>[] = [
   {
