+++ server/src/services/admin/admin.service.ts
@@ -1,5 +1,5 @@
 import { isNil } from 'lodash';
-import { Id, StrapiContext } from '../../@types-v5';
+import { Id, StrapiContext } from '../../@types';
 import { APPROVAL_STATUS } from '../../const';
 import { getCommentRepository, getDefaultAuthorPopulate, getReportCommentRepository } from '../../repositories';
 import { getPluginService } from '../../utils/getPluginService';
