+++ server/src/test/utils.ts
@@ -1,0 +1,1 @@
+export const caster = <T>(source: unknown)=> source as T;
